Emit rinfo to server request handlers
I'm looking to build a DNS server that has conditional behavior based on the source of DNS requests. Currently, the server in this library doesn't surface the remote information to request handlers. This change allows a handler like `server.on('request', (request, send, rinfo) => {})` and will not break any existing interfaces.